Chore: rename GenericWarningModal to WarningModal
