--- src/components/GenericWarningModal/index.ts
@@ -1,1 +1,0 @@
-export * from "./GenericWarningModal"
