+++ src/layouts/EditContactUs.jsx
@@ -6,9 +6,9 @@ import EditorSection from "components/contact-us/Section"
 import FormContext from "components/Form/FormContext"
 import FormTitle from "components/Form/FormTitle"
 import FormField from "components/FormField"
-import { GenericWarningModal } from "components/GenericWarningModal"
 import Header from "components/Header"
 import { LoadingButton } from "components/LoadingButton"
+import { GenericWarningModal } from "components/WarningModal"
 import update from "immutability-helper"
 import _ from "lodash"
 import PropTypes from "prop-types"
