+++ src/layouts/EditHomepage.jsx
@@ -2,7 +2,6 @@
 import { useDisclosure } from "@chakra-ui/react"
 import { Button, Input } from "@opengovsg/design-system-react"
 import axios from "axios"
-import { GenericWarningModal } from "components/GenericWarningModal"
 import Header from "components/Header"
 import EditorHeroSection from "components/homepage/HeroSection"
 import EditorInfobarSection from "components/homepage/InfobarSection"
@@ -10,6 +9,7 @@ import EditorInfopicSection from "components/homepage/InfopicSection"
 import NewSectionCreator from "components/homepage/NewSectionCreator"
 import EditorResourcesSection from "components/homepage/ResourcesSection"
 import { LoadingButton } from "components/LoadingButton"
+import { GenericWarningModal } from "components/WarningModal"
 import update from "immutability-helper"
 import _ from "lodash"
 import PropTypes from "prop-types"
