+++ src/layouts/EditNavBar.jsx
@@ -1,8 +1,8 @@
 import { HStack, useDisclosure } from "@chakra-ui/react"
 import { Button } from "@opengovsg/design-system-react"
-import { GenericWarningModal } from "components/GenericWarningModal"
 import Header from "components/Header"
 import NavSection from "components/navbar/NavSection"
+import { GenericWarningModal } from "components/WarningModal"
 import update from "immutability-helper"
 import _ from "lodash"
 import PropTypes from "prop-types"
