+++ src/layouts/EditPage.jsx
@@ -2,10 +2,10 @@ import { useDisclosure } from "@chakra-ui/react"
 import { Button } from "@opengovsg/design-system-react"
 import axios from "axios"
 import Footer from "components/Footer"
-import { GenericWarningModal } from "components/GenericWarningModal"
 import Header from "components/Header"
 import MarkdownEditor from "components/pages/MarkdownEditor"
 import PagePreview from "components/pages/PagePreview"
+import { GenericWarningModal } from "components/WarningModal"
 import DOMPurify from "dompurify"
 import { marked } from "marked"
 import PropTypes from "prop-types"
